requirements.txt
Flask
mysql-connector-python
Pillow
python-dotenv
requests
tqdm
